changes to enable profiling with serial Flye

+++ strainy/transform.py
@@ -406,7 +406,8 @@ def strong_tail(cluster, cl, ln, data):
 
 
 def gcu_worker(*args):
-
+    if StRainyArgs().threads == 1:
+        args = args[0]
     edge = args[0]
     flye_consensus = args[1]
     strainy_args = args[2]
@@ -445,15 +446,21 @@ def parallelize_gcu(graph_edges, flye_consensus, graph, args):
 
     init_args = [[edge, flye_consensus, args] for edge in graph_edges]
 
-    results = pool.starmap_async(gcu_worker, init_args, chunksize=1)
-    while not results.ready():
-        time.sleep(0.01)
-        if not results._success:
-            pool.terminate()
-            raise Exception("Error in worker thread, exiting")
+    if StRainyArgs().threads == 1:
+        result_values = []
+        for i in init_args:
+            result_values.append(gcu_worker(i))
 
-    pool.close()
-    pool.join()
+    else:
+        results = pool.starmap_async(gcu_worker, init_args, chunksize=1)
+        while not results.ready():
+            time.sleep(0.01)
+            if not results._success:
+                pool.terminate()
+                raise Exception("Error in worker thread, exiting")
+        result_values = results._value
+        pool.close()
+        pool.join()
 
     bam_cache = {}
     link_clusters = defaultdict(list)
@@ -464,7 +471,7 @@ def parallelize_gcu(graph_edges, flye_consensus, graph, args):
     
     outputs = [bam_cache, link_clusters, link_clusters_src, link_clusters_sink, graph_ops, remove_clusters]
     # join the results of multiple threads
-    for r in results._value:
+    for r in result_values:
         for i in range(len(r)):
             if i == len(r) - 1 or i == len(r) - 2:
                 for k in r[i]:
@@ -858,17 +865,7 @@ def transform_main(args):
 
     flye_consensus = FlyeConsensus(StRainyArgs().bam, StRainyArgs().fa, args.threads, consensus_dict, multiprocessing.Manager())
 
-    # bam_cache = {}
-    # link_clusters = defaultdict(list)
-    # link_clusters_src = defaultdict(list)
-    # link_clusters_sink = defaultdict(list)
-    # remove_clusters = set()
-
     logger.info("### Create unitigs")
-    # for edge in StRainyArgs().edges:
-    #     #TODO: this can run in parallel (and probably takes the most time)
-    #     graph_create_unitigs(edge, initial_graph, flye_consensus, bam_cache,
-    #                          link_clusters, link_clusters_src, link_clusters_sink, remove_clusters)
 
     bam_cache, link_clusters, link_clusters_src, link_clusters_sink, remove_clusters, initial_graph = parallelize_gcu(StRainyArgs().edges, flye_consensus, initial_graph, args)
 
@@ -877,7 +874,6 @@ def transform_main(args):
         graph_link_unitigs(edge, initial_graph, bam_cache, link_clusters, link_clusters_src,
                            link_clusters_sink, remove_clusters)
     connect_parental_edges(initial_graph, link_clusters_src, link_clusters_sink, remove_clusters)
-    #gfapy.Gfa.to_file(initial_graph, StRainyArgs().gfa_transformed)
 
     logger.info("### Remove initial segments")
     for ed in initial_graph.segments:
